tema4/requirements.txt
Flask==3.1.0
ms-identity-python @ git+https://github.com/azure-samples/ms-identity-python@834583646df315b79bc89f54808f4ae9c2a1ce59
python-dotenv==1.1.0
requests==2.32.3
flask_session==0.8.0
